Version.1.4.14
アイテム
＊胴当てMark.1
＊胴当てMark.2
＊胴当てMark.3
＊胴当てMark.4

+++ src/main/java/ayato/entity/Enemy.java
@@ -10,7 +10,7 @@ import java.util.Random;
 public class Enemy extends AbstractEntity{
     JsonNode eStates;
     private final int U_HP, U_MP, U_G, U_ATK, U_EXP;
-    private final double U_DEFENCE;
+    private final int U_DEFENCE;
     public Enemy(LunchScene m, ImageMaker maker, JsonNode entity_states) {
         super(m, maker, 0, 0, 0, 0);
         eStates = entity_states;
@@ -19,7 +19,7 @@ public class Enemy extends AbstractEntity{
         U_ATK = entity_states.get(EnemyFactory.U_ATK).asInt();
         U_EXP = entity_states.get(EnemyFactory.U_EXP).asInt();
         U_MP = entity_states.get(EnemyFactory.U_MP).asInt();
-        U_DEFENCE = entity_states.get(EnemyFactory.U_DF).asDouble();
+        U_DEFENCE = entity_states.get(EnemyFactory.U_DF).asInt();
         reset();
     }
     public void setEnemyLevel(int playerLv, int minLv,  int maxLv){
@@ -48,7 +48,7 @@ public class Enemy extends AbstractEntity{
                 eStates.get(EnemyFactory.G).asInt(),
                 eStates.get(EnemyFactory.ATK).asInt(),
                 eStates.get(EnemyFactory.POW).asInt(),
-                eStates.get(EnemyFactory.DF).asDouble()
+                eStates.get(EnemyFactory.DF).asInt()
                 );
     }
 }
